Let caddy handle redirecting requests to dev/prod
Instead of having different URLs for dev and prod, and having to trust
different origins, unify everything to be like it would be in prod by
using Caddy to detect which mode is being used and handling all the
redirections. This makes for a transparent experience, with fewer
headaches and complications.